+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v0.0.2
+
+### Added or Changed
+- Added create project
+- Added delete project
+
 ## v0.0.1
 
 ### Added or Changed